finish support to gin on glf with dgl backend
